Terrain inherits from base class + efficient rendering

+++ src/terrain/BaseTerrain.java
@@ -18,6 +18,6 @@ public abstract class BaseTerrain {
 	}
 	
 	public abstract double[][] generate();
-	public abstract double[][] generate(int[] arguments);
+	public abstract double[][] generate(long[] arguments);
 	
 }
